requirements.txt
streamlit
pandas
scikit-learn
tweepy
google-search-results
anthropic
resend
apscheduler
python-dotenv
plotly